# api_spec.json
{"openapi":"3.1.0","info":{"title":"AI-Powered Recommendation Engine","description":"API for the AI-Powered Recommendation Engine.","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register User","operationId":"register_user_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenWithAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenWithAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/users/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_v1_auth_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/forgot-password":{"post":{"tags":["auth"],"summary":"Forgot Password","description":"Send password reset token to user's email","operationId":"forgot_password_api_v1_auth_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/reset-password":{"post":{"tags":["auth"],"summary":"Reset Password","description":"Reset user password using reset token","operationId":"reset_password_api_v1_auth_reset_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/change-password":{"post":{"tags":["auth"],"summary":"Change Password","description":"Change user password (requires current password)","operationId":"change_password_api_v1_auth_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/assessment/upload-resume":{"post":{"tags":["assessment"],"summary":"Upload Resume","operationId":"upload_resume_api_v1_assessment_upload_resume_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_api_v1_assessment_upload_resume_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/assessment/submit-assessment":{"post":{"tags":["assessment"],"summary":"Submit Assessment","operationId":"submit_assessment_api_v1_assessment_submit_assessment_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Assessment Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/assessment/get-latest-assessment":{"get":{"tags":["assessment"],"summary":"Get Latest Assessment","operationId":"get_latest_assessment_api_v1_assessment_get_latest_assessment_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/recommendation/generate-recommendations":{"post":{"tags":["recommendation"],"summary":"Generate Recommendations","operationId":"generate_recommendations_api_v1_recommendation_generate_recommendations_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__models__recommendation__RecommendationSet"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/recommendation/tune-recommendations":{"post":{"tags":["recommendation"],"summary":"Tune Recommendations","operationId":"tune_recommendations_api_v1_recommendation_tune_recommendations_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TuneRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__models__recommendation__RecommendationSet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/recommendation/get-latest-recommendations":{"get":{"tags":["recommendation"],"summary":"Get Latest Recommendations","operationId":"get_latest_recommendations_api_v1_recommendation_get_latest_recommendations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__models__recommendation__RecommendationSet"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/recommendations":{"post":{"tags":["ai"],"summary":"Get Ai Recommendations","operationId":"get_ai_recommendations_api_v1_ai_recommendations_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAssessment"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__models__ai__RecommendationSet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/tune":{"post":{"tags":["ai"],"summary":"Tune Ai Recommendations","operationId":"tune_ai_recommendations_api_v1_ai_tune_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TuningPrompt"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__models__ai__RecommendationSet"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/chat/":{"post":{"tags":["chat"],"summary":"Chat With Ai","description":"Handles chat messages from the user and returns an AI-generated response.","operationId":"chat_with_ai_api_v1_chat__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v1_auth_login_post"},"Body_upload_resume_api_v1_assessment_upload_resume_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_resume_api_v1_assessment_upload_resume_post"},"ChangePasswordRequest":{"properties":{"current_password":{"type":"string","title":"Current Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["current_password","new_password"],"title":"ChangePasswordRequest"},"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"reply":{"type":"string","title":"Reply"}},"type":"object","required":["reply"],"title":"ChatResponse"},"ForgotPasswordRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"ForgotPasswordRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobRecommendation":{"properties":{"job_id":{"type":"string","title":"Job Id"},"title":{"type":"string","title":"Title"},"seniority":{"type":"string","title":"Seniority"},"location":{"type":"string","title":"Location"},"score":{"type":"number","title":"Score"},"highlights":{"items":{"type":"string"},"type":"array","title":"Highlights"},"role":{"type":"string","title":"Role"},"tech":{"items":{"type":"string"},"type":"array","title":"Tech"},"employment_type":{"type":"string","title":"Employment Type"},"industry":{"type":"string","title":"Industry"}},"type":"object","required":["job_id","title","seniority","location","score","highlights","role","tech","employment_type","industry"],"title":"JobRecommendation"},"Recommendation":{"properties":{"job_id":{"type":"string","title":"Job Id"},"title":{"type":"string","title":"Title"},"company":{"type":"string","title":"Company"},"location":{"type":"string","title":"Location"},"description":{"type":"string","title":"Description"},"requirements":{"items":{"type":"string"},"type":"array","title":"Requirements"},"seniority":{"type":"string","title":"Seniority"},"score":{"type":"number","title":"Score"},"highlights":{"items":{"type":"string"},"type":"array","title":"Highlights"},"role":{"type":"string","title":"Role"},"tech":{"items":{"type":"string"},"type":"array","title":"Tech"},"employment_type":{"type":"string","title":"Employment Type"},"industry":{"type":"string","title":"Industry"}},"type":"object","required":["job_id","title","company","location","description","requirements","seniority","score","highlights","role","tech","employment_type","industry"],"title":"Recommendation"},"ResetPasswordRequest":{"properties":{"token":{"type":"string","title":"Token"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["token","new_password"],"title":"ResetPasswordRequest"},"TokenWithAssessment":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"assessment_completed":{"type":"boolean","title":"Assessment Completed"}},"type":"object","required":["access_token","token_type","assessment_completed"],"title":"TokenWithAssessment"},"TuneRequest":{"properties":{"recommendation_set_id":{"type":"string","title":"Recommendation Set Id"},"prompt":{"type":"string","title":"Prompt"}},"type":"object","required":["recommendation_set_id","prompt"],"title":"TuneRequest"},"TuningPrompt":{"properties":{"current_recommendations":{"items":{"type":"object"},"type":"array","title":"Current Recommendations"},"prompt":{"type":"string","title":"Prompt"}},"type":"object","required":["current_recommendations","prompt"],"title":"TuningPrompt"},"User":{"properties":{"_id":{"type":"string","title":" Id"},"email":{"type":"string","format":"email","title":"Email"},"hashed_password":{"type":"string","title":"Hashed Password"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","hashed_password"],"title":"User"},"UserAssessment":{"properties":{"_id":{"type":"string","title":" Id"},"user_id":{"type":"string","title":"User Id"},"age":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Age"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"educationLevel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Educationlevel"},"currentSituation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Currentsituation"},"currentRole":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Currentrole"},"experience":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Experience"},"resumeText":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resumetext"},"linkedinProfile":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linkedinprofile"},"certifications":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Certifications","default":[]},"otherTradesCert":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Othertradescert"},"otherTechCert":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Othertechcert"},"otherBusinessCert":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Otherbusinesscert"},"otherHealthcareCert":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Otherhealthcarecert"},"technicalSkills":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Technicalskills","default":[]},"softSkills":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Softskills","default":[]},"workingWithData":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Workingwithdata","default":3},"workingWithPeople":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Workingwithpeople","default":3},"creativeTasks":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Creativetasks","default":3},"problemSolving":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Problemsolving","default":3},"leadership":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Leadership","default":3},"physicalHandsOnWork":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Physicalhandsonwork","default":3},"outdoorWork":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Outdoorwork","default":3},"mechanicalAptitude":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Mechanicalaptitude","default":3},"interests":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Interests","default":[]},"industries":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Industries","default":[]},"workEnvironment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Workenvironment"},"careerGoals":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Careergoals"},"workLifeBalance":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Worklifebalance"},"salaryExpectations":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Salaryexpectations"},"skills":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Skills"},"career_goals":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Career Goals"},"preferences":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Preferences"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["user_id"],"title":"UserAssessment"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"app__models__ai__RecommendationSet":{"properties":{"recommendations":{"items":{"$ref":"#/components/schemas/Recommendation"},"type":"array","title":"Recommendations"}},"type":"object","required":["recommendations"],"title":"RecommendationSet"},"app__models__recommendation__RecommendationSet":{"properties":{"_id":{"type":"string","title":" Id"},"user_id":{"type":"string","title":"User Id"},"recommendations":{"items":{"$ref":"#/components/schemas/JobRecommendation"},"type":"array","title":"Recommendations"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["user_id","recommendations"],"title":"RecommendationSet"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}